Modularize application
