+++ src/config.ts
@@ -1,0 +1,11 @@
+export interface OverlayConfig {
+  duration: number;
+  opacity: number;
+  font: string;
+}
+
+export const overlayConfig: OverlayConfig = {
+  duration: 500,
+  opacity: 0.8,
+  font: 'Arial'
+};
